requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
